Moved types common to server-side and smart data source to a shared file.
