+++ source/components/cardContainer/dataSources/serverSideDataSource/serverSideDataSource.service.ts
@@ -5,37 +5,13 @@ import __array = services.array;
 import __object = services.object;
 import __synchronizedRequests = services.synchronizedRequests;
 
+import { IServerSearchFunction, IServerSearchParams, ISortParams, IPagingParams, IDataResult } from '../asyncTypes';
 import { IAsyncDataSource, AsyncDataSource, IDataSetFunction } from '../asyncDataSource.service';
 import { IDataSourceProcessor, processorServiceName } from '../dataSourceProcessor.service';
 import { ISort, SortDirection } from '../../sorts/sort';
 
 export interface IServerSideDataSource<TDataType> extends IAsyncDataSource<TDataType> {
 	filters: filters.ISerializableFilter<any>[];
-}
-
-export interface IServerSearchFunction<TDataType> {
-	(searchParams: IServerSearchParams): angular.IPromise<IDataResult<TDataType>>;
-}
-
-export interface IServerSearchParams {
-	filters: {[index: string]: any};
-	sorts: ISortParams[];
-	paging: IPagingParams;
-}
-
-export interface ISortParams {
-	column: string;
-	direction: string;
-}
-
-export interface IPagingParams {
-	pageNumber: number;
-	pageSize: number;
-}
-
-export interface IDataResult<TDataType> {
-	dataSet: TDataType[];
-	count: number;
 }
 
 export class ServerSideDataSource<TDataType> extends AsyncDataSource<TDataType> {
